+++ release_notes.md
@@ -1,3 +1,7 @@
+## NEXT
+* use more standard name for prefetch configuration
+* override default configuration with queue configuration
+
 ## 2.0.1
 * various bug fixes